fixed unnatural ball-paddle bounce
If changing both X and Y direction of the ball is changed, the ball in fact
goes backwards. In order to correctly handle bounces with the paddle, we
should take care of which border of the paddle was hit by the ball. To keep
the code simple, we can ignore side bounces and let the ball bounce only
horizontally.